requirements.txt
# Core Framework
streamlit>=1.28.0
python-dotenv>=1.0.0

# Groq (OpenAI-compatible client)
openai>=1.35.0  # Needed for Groq’s /openai/v1 endpoint
requests>=2.31.0

# Ollama (local fallback)
ollama>=0.1.0

# LangChain Ecosystem
langchain>=0.1.0
langchain-community>=0.0.20
langchain-huggingface>=0.0.3
langchain-text-splitters>=0.0.1

# Vector Stores
faiss-cpu>=1.7.4
chromadb>=0.4.0

# Embeddings & ML
sentence-transformers>=2.2.2
huggingface-hub>=0.19.0
tiktoken>=0.5.0

# Document Processing
PyPDF2>=3.0.0
pypdf>=3.17.0
PyMuPDF>=1.23.0
python-docx>=1.1.0
docx2txt>=0.8

# Image & Media
Pillow>=10.0.0
